Compute the wildfire exposure metric with custom transmission distances

For advanced users. If the transmission distances from the wildfire exposure literature are not representative of the wildland fuels in your area of interest, this function can be used to change the transmission distance to a custom distance. It is highly recommended that any exposure layers produced with this function are validated with observed fire history using the fire_exp_validate() function.

Usage

fire_exp_adjust(hazard, tdist, nonburnable)

Arguments

hazard

a SpatRaster that represents hazardous fuels for the transmission distance specified in tdist

tdist

Numeric, transmission distance in meters

nonburnable

(Optional) SpatRaster that represents the non-burnable landscape. Any cells that cannot receive wildfire (e.g. open water, rock) and any cells that are not natural (e.g. built environment, irrigated agricultural areas) should be of value 1, all other cells should be NODATA. This parameter should be provided if preparing data for fire_exp_validate()

Value

SpatRaster object of exposure values

Examples

# read example hazard data ----------------------------------
filepath <- "extdata/hazard.tif"
haz <- terra::rast(system.file(filepath, package = "fireexposuR"))
# -----------------------------------------------------------

# compute long range exposure with custom disance of 800 m
exp <- fire_exp_adjust(haz, tdist = 800)