fireexposuR: Compute and Visualize Wildfire Exposure

This package computes and visualizes wildfire exposure using the methods documented in a series of scientific publications.

See also

Author

Maintainer: Air Forbes amforbes@ualberta.ca (ORCID)

Other contributors:

Jennifer Beverly [thesis advisor]