Compute the wildfire exposure metric

fire_exp() computes the wildfire exposure metric from a hazard fuel raster. The hazard fuel raster must be prepared by the user. Forbes and Beverly 2024 (manuscript in preparation) details suggestions for data acquisition and preparation in accordance with various budget limitations and user experience levels. Computing wildfire exposure for different transmission distances may require unique hazard rasters for each distance. See Beverly et al. 2021, Beverly et al. 2010, and Schmidt et al. 2024 for further references for hazard raster preparation.

Usage

fire_exp(hazard, tdist = c("l", "s", "r"), nonburnable)

Arguments

hazard

a SpatRaster that represents hazardous fuels for the transmission distance specified in tdist

tdist

a character vector, can be: "l" for long-range embers (Default), "s" for short-range embers or, "r" for radiant heat

nonburnable

(optional) a SpatRaster that represents the non-burnable landscape. Any cells that cannot receive wildfire (e.g. open water, rock) and any cells that are not natural (e.g. built environment, irrigated agricultural areas) should be of value 1, all other cells should be NODATA. This parameter should be provided if preparing data for fire_exp_validate()

Value

A SpatRaster object of exposure values between 0-1

Details

There are minimum raster resolution requirements for each transmission distance. Radiant heat requires data of 10 m or finer, short-range embers 90 m or finer, and long-range embers requires 150 m or finer.

Examples

# read example hazard data ----------------------------------
filepath <- "extdata/hazard.tif"
haz <- terra::rast(system.file(filepath, package = "fireexposuR"))
# -----------------------------------------------------------

# compute long range exposure
exp <- fire_exp(haz, tdist = "l")
#> Warning: RGL: unable to open X11 display
#> Warning: 'rgl.init' failed, running with 'rgl.useNULL = TRUE'.
exp
#> class       : SpatRaster 
#> dimensions  : 1000, 1000, 1  (nrow, ncol, nlyr)
#> resolution  : 100, 100  (x, y)
#> extent      : 450000, 550000, 4950000, 5050000  (xmin, xmax, ymin, ymax)
#> coord. ref. : WGS 84 / UTM zone 8N (EPSG:32608) 
#> source(s)   : memory
#> varname     : hazard 
#> name        : exposure 
#> min value   :        0 
#> max value   :        1 

# each transmission distance has a resolution requirement and exposure() will
# not run if resolution is too coarse
try(fire_exp(haz, tdist = "r"))
#> Error in fire_exp(haz, tdist = "r") : 
#>   Insufficient resolution for radiant heat exposure assessment