Generate a directional load plot or table for multiple values

fire_exp_dir_multi() automates the directional vulnerability assessment methods from Beverly and Forbes 2023. This function can return directional loads as:

a standardized radial plot as a ggplot object

a table summarizing if each degree is included by feature

Usage

fire_exp_dir_multi(exposure, values, plot = FALSE, all = FALSE)

Arguments

exposure

SpatRaster from fire_exp()

values

Spatvector of value as a point or simplified polygon

plot

Boolean, when TRUE: returns a standardized directional plot. The default is FALSE.

all

Boolean, when TRUE: considers all 3 segments (0-15km) of directional transects. when FALSE: only the segments from 5-15 km are included (Default)

Value

a data.frame of the features with attributes: value featureID, degree, to5 (binary), to10(binary), t015(binary), full(binary), outer (binary). Unless: * plot = TRUE: a standardized plot as a ggplot object

Examples

#' # read example hazard data ----------------------------------
filepath <- "extdata/hazard.tif"
haz <- terra::rast(system.file(filepath, package = "fireexposuR"))
# example points across the landscape
e <- terra::buffer(terra::vect(terra::ext(haz), crs = haz), -15500)
pts <- terra::spatSample(e, 20)
# ----------------------------------------------------------

exp <- fire_exp(haz, tdist = "l")
# this example will take a while to run
if (FALSE) { # \dontrun{
fire_exp_dir_multi(exp, pts, plot = TRUE)
} # }