Summarize exposure by class

fire_exp_summary() creates a summary table of area and proportions of exposure in predetermined exposure classes.

Usage

fire_exp_summary(exposure, aoi, classify = c("landscape", "local"))

Arguments

exposure

SpatRaster from fire_exp()

aoi

(optional) SpatVector of an area of interest to mask exposure for summary

classify

character, either "landscape" or "local". default is "landscape".

Value

a summary table as a data frame object

Details

DOCUMENTATION IN DEVELOPMENT

The table reports the number of pixels, the proportion, and area in hectares and meters squared by class.

Landscape classification breaks are:

Low (0-20%)

Moderate (20-40%)

High (40-60%),

Very High (60-80%)

Extreme (80-100%)

Local classification breaks are:

Nil (0%)

Low (>0-15%)

Moderate (15-30%)

High (30-45%)

Extreme (45%+)

Examples

# read example hazard data
hazard_file_path <- "extdata/hazard.tif"
hazard <- terra::rast(system.file(hazard_file_path, package = "fireexposuR"))

# read example area of interest polygon geometry
geom_file_path <- "extdata/polygon_geometry.csv"
geom <- read.csv(system.file(geom_file_path, package = "fireexposuR"))
aoi <- terra::vect(as.matrix(geom), "polygons", crs = hazard)

# Compute exposure
exposure <- fire_exp(hazard)

# Summary for full extent of data
fire_exp_summary(exposure, classify = "landscape")
#>       class npixels      prop    aream2 areaha
#> 1       Low   14187 0.1466796 141870000  14187
#> 2  Moderate   13073 0.1351620 130730000  13073
#> 3      High   15105 0.1561708 151050000  15105
#> 4 Very High   18911 0.1955211 189110000  18911
#> 5   Extreme   35445 0.3664664 354450000  35445

# Summary masked to an area of interest
fire_exp_summary(exposure, aoi, classify = "landscape")
#>       class npixels       prop  aream2 areaha
#> 1       Low      57 0.06761566  570000     57
#> 2  Moderate     172 0.20403321 1720000    172
#> 3      High     207 0.24555160 2070000    207
#> 4 Very High     330 0.39145907 3300000    330
#> 5   Extreme      77 0.09134045  770000     77