Summarize exposure by class

fire_exp_summary() creates a summary table of exposure by the specified classification scheme. The table reports the number of pixels, the proportion, and area in hectares and meters squared by class.

Landscape classification breaks are:

Low (0-20%)

Moderate (20-40%)

High (40-60%),

Very High (60-80%)

Extreme (80-100%)

Local classification breaks are:

Nil (0%)

Low (>0-15%)

Moderate (15-30%)

High (30-45%)

Extreme (45%+)

Usage

fire_exp_summary(exposure, aoi, classify = c("landscape", "local"))

Arguments

exposure

SpatRaster from fire_exp()

aoi

(optional) SpatVector of an area of interest to mask exposure for summary

classify

character, either "landscape" or "local". default is "landscape".

Value

a summary table as a data frame object

Examples

#' # read example hazard data ----------------------------------
filepath <- "extdata/hazard.tif"
haz <- terra::rast(system.file(filepath, package = "fireexposuR"))
# generate example AOI polygon -----------------------------
filepath <- "extdata/builtsimpleexamplegeom.csv"
g <- read.csv(system.file(filepath, package = "fireexposuR"))
v <- terra::vect(as.matrix(g), "polygons", crs = haz)
# ----------------------------------------------------------

exp <- fire_exp(haz)

# for full extent of data
fire_exp_summary(exp, classify = "landscape")
#>       class npixels      prop     aream2 areaha
#> 1       Low  320073 0.3265718 3200730000 320073
#> 2  Moderate  118269 0.1206703 1182690000 118269
#> 3      High  121896 0.1243710 1218960000 121896
#> 4 Very High  115884 0.1182369 1158840000 115884
#> 5   Extreme  303978 0.3101500 3039780000 303978

# for a masked area
fire_exp_summary(exp, v, classify = "landscape")
#>       class npixels        prop  aream2 areaha
#> 1       Low     703 0.835909631 7030000    703
#> 2  Moderate      73 0.086801427  730000     73
#> 3      High      62 0.073721760  620000     62
#> 4 Very High       3 0.003567182   30000      3